refactor(http-bridge): Re-home HTTP concrete bridges
Move the remaining concrete HTTP bridge implementations out of runtime.endpoints.http* and into adapter-owned clients.http.bridge* packages.

Keep HttpShellBridgeFactories as the default production binding entry point, update bootstrap and admin wiring, and move the affected tests with the bridge packages so package-private visibility stays narrow.
