+++ src/main/java/network/crypta/runtime/endpoints/http/HttpShellBridgeFactories.java
@@ -1,16 +1,18 @@
 package network.crypta.runtime.endpoints.http;
 
+import network.crypta.clients.http.bridge.CoreHttpShellRuntimeSupport;
+import network.crypta.clients.http.bridge.HttpShellContainers;
 import network.crypta.runtime.http.HttpShellContainerFactory;
 import network.crypta.runtime.http.HttpShellRuntimeSupportFactory;
 
 /**
  * Endpoint-owned default production bindings for the runtime HTTP shell seams.
  *
- * <p>This helper keeps knowledge of the current bridge implementations inside the endpoint-owned
- * package while letting higher-level runtime/bootstrap code depend only on the neutral seam types
- * from {@code network.crypta.runtime.http}. Bootstrap code calls these accessors when it wants the
- * historical daemon wiring, but the neutral seam package stays free of references back to
- * endpoint-owned bridge classes. That separation keeps ownership boundaries clear without changing
+ * <p>This helper keeps the default production binding choice inside the runtime endpoint package
+ * while letting higher-level runtime/bootstrap code depend only on the neutral seam types from
+ * {@code network.crypta.runtime.http}. Bootstrap code calls these accessors when it wants the
+ * historical daemon wiring, but the neutral seam package stays free of references back to the
+ * adapter-owned bridge classes. That separation keeps ownership boundaries clear without changing
  * startup order, shell lifecycle sequencing, or the current bridge implementations used in
  * production.
  *
@@ -31,7 +33,7 @@ public final class HttpShellBridgeFactories {
    * when the node services subsystem is ready to construct the HTTP shell for a specific
    * configuration and executor.
    *
-   * @return factory that creates the current endpoint-backed HTTP shell container bridge without
+   * @return factory that creates the current adapter-backed HTTP shell container bridge without
    *     eagerly constructing the shell
    */
   public static HttpShellContainerFactory defaultContainerFactory() {
@@ -49,8 +51,8 @@ public final class HttpShellBridgeFactories {
    * with both the runtime-owned seam and the legacy HTTP shell adapter, so they match the default
    * container factory returned by {@link #defaultContainerFactory()}.
    *
-   * @return factory that creates the current endpoint-backed HTTP shell runtime support bridge for
-   *     a supplied daemon core
+   * @return factory that creates the current adapter-backed HTTP shell runtime support bridge for a
+   *     supplied daemon core
    */
   public static HttpShellRuntimeSupportFactory coreBackedRuntimeSupportFactory() {
     return CoreHttpShellRuntimeSupport::new;
