+++ src/main/java/network/crypta/runtime/endpoints/http/package-info.java
@@ -1,11 +1,10 @@
 /**
- * Endpoint-owned HTTP bridge implementations and default production bindings.
+ * Endpoint-owned default production bindings for HTTP bridge implementations.
  *
- * <p>This package adapts daemon- and runtime-backed services to the narrow runtime-support seams
- * owned by {@code network.crypta.clients.http}. It contains the concrete HTTP shell bridge
- * implementations plus the endpoint-owned helper methods that expose the default production
- * bindings used by bootstrap. Higher-level runtime and bootstrap code should depend on the neutral
- * seam types in {@code network.crypta.runtime.http}, not on the bridge implementation classes in
- * this package.
+ * <p>This package keeps the bootstrap-facing helper methods that expose the default production
+ * bindings used by runtime/bootstrap code. The concrete HTTP bridge implementations now live under
+ * {@code network.crypta.clients.http.bridge}. Higher-level runtime and bootstrap code should depend
+ * on the neutral seam types in {@code network.crypta.runtime.http}, not on the bridge
+ * implementation classes themselves.
  */
 package network.crypta.runtime.endpoints.http;
