+++ src/test/java/network/crypta/clients/http/SimpleToadletServerTest.java
@@ -9,6 +9,8 @@ import network.crypta.client.FetchContext;
 import network.crypta.client.HighLevelSimpleClient;
 import network.crypta.client.async.ClientContext;
 import network.crypta.clients.http.bookmark.BookmarkManager;
+import network.crypta.clients.http.bridge.CoreHttpShellRuntimeSupport;
+import network.crypta.clients.http.bridge.bookmark.CoreBookmarkRuntimeSupport;
 import network.crypta.config.Config;
 import network.crypta.config.PersistentConfig;
 import network.crypta.config.SubConfig;
@@ -17,8 +19,6 @@ import network.crypta.io.SSLNetworkInterface;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.RequestStarter;
 import network.crypta.runtime.alerts.UserAlertManager;
-import network.crypta.runtime.endpoints.http.CoreHttpShellRuntimeSupport;
-import network.crypta.runtime.endpoints.http.bookmark.CoreBookmarkRuntimeSupport;
 import network.crypta.runtime.spi.RandomnessPort;
 import network.crypta.runtime.spi.RuntimePorts;
 import network.crypta.support.HTMLNode;
