+++ src/test/java/network/crypta/runtime/bootstrap/NodeRuntimeBridgeFactoriesTest.java
@@ -1,6 +1,9 @@
 package network.crypta.runtime.bootstrap;
 
 import java.io.File;
+import network.crypta.clients.http.bridge.CoreHttpShellRuntimeSupport;
+import network.crypta.clients.http.bridge.geoip.HttpGeoIpCountryLookup;
+import network.crypta.clients.http.bridge.security.CorePasswordFormPageRenderer;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.ProgramDirectory;
@@ -8,9 +11,6 @@ import network.crypta.runtime.admin.AdminRuntimeBridgeInputs;
 import network.crypta.runtime.admin.AdminRuntimeBridgeInputsFactory;
 import network.crypta.runtime.endpoints.fcp.FcpPersistentRequestServices;
 import network.crypta.runtime.endpoints.fcp.FcpQueueAdminBackend;
-import network.crypta.runtime.endpoints.http.CoreHttpShellRuntimeSupport;
-import network.crypta.runtime.endpoints.http.geoip.HttpGeoIpCountryLookup;
-import network.crypta.runtime.endpoints.http.security.CorePasswordFormPageRenderer;
 import network.crypta.runtime.fcp.PersistentRequestEndpointServices;
 import network.crypta.runtime.fcp.PersistentRequestEndpointServicesFactory;
 import network.crypta.runtime.http.HttpShellContainerFactory;
